Split VM tests
